+++ trunk/src/com/geoscope/GeoEye/Space/TypesSystem/Positioner/TTPositionerFunctionality.java
@@ -21,7 +21,7 @@ public class TTPositionerFunctionality extends TTypeFunctionality {
 	}
 	
 	@Override
-	public TComponentFunctionality TComponentFunctionality_Create(int idComponent) {
+	public TComponentFunctionality TComponentFunctionality_Create(long idComponent) {
 		return (new TPositionerFunctionality(this,idComponent));
 	}
 }
